requirements-dev.txt
# Include production requirements
-r requirements.txt

# Development dependencies
black==23.12.1
flake8==7.0.0
isort==5.13.2
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
python-dotenv==1.0.0
